refactor: move initial value tests to their respective test classes
